#16: Move current vitamins into a vitamin package
