+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/CenterOfMass.kt
@@ -14,43 +14,10 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin
 
-/**
- * A generic stepper motor.
- *
- * Good things to put in [Vitamin.specs]:
- *  - Mounting holes (Vitamin for the screws and hole positions)
- */
-interface StepperMotor : Vitamin {
-
-    /**
-     * The shaft diameter.
-     */
-    val shaftDiameter: Double
-
-    /**
-     * The NEMA size.
-     */
-    val nemaSize: Int
-
-    /**
-     * The nominal voltage.
-     */
-    val voltage: Double
-
-    /**
-     * The holding torque.
-     */
-    val holdingTorque: Double
-
-    /**
-     * The rated current.
-     */
-    val current: Double
-
-    /**
-     * The degrees per step.
-     */
-    val stepAngle: Double
-}
+data class CenterOfMass(
+    val x: Double,
+    val y: Double,
+    val z: Double
+)
