+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/DefaultDCMotor.kt
@@ -14,15 +14,20 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin
 
 import com.google.common.collect.ImmutableMap
 import com.neuronrobotics.bowlerkernel.gitfs.GitFile
+import com.neuronrobotics.bowlerkernel.util.Limits
 
-data class DefaultServo(
-    override val voltage: Double,
+data class DefaultDCMotor(
+    override val voltage: Limits,
+    override val outputShaftDiameter: Double,
+    override val freeSpeed: Double,
+    override val freeCurrent: Double,
     override val stallTorque: Double,
-    override val speed: Double,
+    override val stallCurrent: Double,
+    override val power: Double,
     override val width: Double,
     override val length: Double,
     override val height: Double,
@@ -30,4 +35,4 @@ data class DefaultServo(
     override val centerOfMass: CenterOfMass,
     override val specs: ImmutableMap<String, Any>,
     override val cadGenerator: GitFile
-) : Servo
+) : DCMotor
