+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/DefaultStepperMotor.kt
@@ -14,16 +14,18 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin
 
 import com.google.common.collect.ImmutableMap
 import com.neuronrobotics.bowlerkernel.gitfs.GitFile
 
-data class DefaultBattery(
+data class DefaultStepperMotor(
+    override val shaftDiameter: Double,
+    override val nemaSize: Int,
     override val voltage: Double,
+    override val holdingTorque: Double,
     override val current: Double,
-    override val dischargeRate: Double,
-    override val capacity: Double,
+    override val stepAngle: Double,
     override val width: Double,
     override val length: Double,
     override val height: Double,
@@ -31,4 +33,4 @@ data class DefaultBattery(
     override val centerOfMass: CenterOfMass,
     override val specs: ImmutableMap<String, Any>,
     override val cadGenerator: GitFile
-) : Battery
+) : StepperMotor
