+++ bowler-kernel/vitamins/src/main/kotlin/com/neuronrobotics/bowlerkernel/vitamins/vitamin/StepperMotor.kt
@@ -14,18 +14,43 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.neuronrobotics.bowlerkernel.vitamins
+package com.neuronrobotics.bowlerkernel.vitamins.vitamin
 
-import com.google.common.collect.ImmutableMap
-import com.neuronrobotics.bowlerkernel.gitfs.GitFile
+/**
+ * A generic stepper motor.
+ *
+ * Good things to put in [Vitamin.specs]:
+ *  - Mounting holes (Vitamin for the screws and hole positions)
+ */
+interface StepperMotor : Vitamin {
 
-data class DefaultBallBearing(
-    override val bore: Double,
-    override val width: Double,
-    override val length: Double,
-    override val height: Double,
-    override val weight: Double,
-    override val centerOfMass: CenterOfMass,
-    override val specs: ImmutableMap<String, Any>,
-    override val cadGenerator: GitFile
-) : BallBearing
+    /**
+     * The shaft diameter.
+     */
+    val shaftDiameter: Double
+
+    /**
+     * The NEMA size.
+     */
+    val nemaSize: Int
+
+    /**
+     * The nominal voltage.
+     */
+    val voltage: Double
+
+    /**
+     * The holding torque.
+     */
+    val holdingTorque: Double
+
+    /**
+     * The rated current.
+     */
+    val current: Double
+
+    /**
+     * The degrees per step.
+     */
+    val stepAngle: Double
+}
